requirements.txt
pymongo
certifi
python-dotenv
leetscrape
beautifulsoup4
requests
tqdm

